Production harden the pipeline error path and exhaustively test.
Port the pipeline teardown code from express-processimage with the
necessary adjustments to make it work correctly with a duplex stream.
As of this commit an error either at the top-level(e.g. coming from
the stream of data being piped in) or from one of the streams that
form the pipeline will cause the entire pipeline to be cleaned up.

This commit also bumps unexpected-stream to 4.0.0 which provides
a couple of useful extra assertions but it up to date.