Handle engine execute() methods throwing after streaming begins.
Streaming in the usual case is triggers by flush() being called once
someone starts piping data into the Pipeline, which itself is duplex
stream. Given there are plain throws in flush(), this meant those
methods could themselves throw when called from _read/_write.

Note that streams expect errors to be communicated by emitting an
error - in order to reconcile these, wrap _flush in a try/catch and
if an errors occurs in a situation where streaming was occurring we
instead opt to call _fail() which should initiate stream teardown.