Exercise a gm stream error and fix the codepath after 554797e.
Arrange a test that triggers no output from gm (due to being an
unsupported image type).

This caused a bug to become apparent - the "finish" event was marking
the stream ended, but not that a Pipeline is a duplex stream and that
event is only one side of that stream. Removing those lines allows
the stream error handling to kick in correctly.