Address an issue in the core Pipeline if all operations were invalid.
When the core operation interface was used there was a care where an
operation list containing no valid operations would caused the list
of candidate engines to be empty. If flush was called in this case
it would assume the array had been set causing a dereference. Add
protection for this case and validate the message that is thrown.