Relocate object based engine disabling syntax to create pipeline.
The Impro constructor is called before any engines are registered
with it. In practice that means that if you wish to exclude an
engine, it is best done by simply not adding it with .use().

When creating a conversion Pipeline however it could be very useful
to disable a previously registered engine for the duration of a set
of operations. Thus allow createPipeline({<engineName>: false}) as
a valid way to disable engines for that pipeline.