Remove root from the set of supported top-level options.
The only engine that uses this value is SvgFilter.. but: there is no
code at the point where the execute method is called that copies the
value from instance options objects to the those passed into to the
engine. Instead, the expectation is that is set via the engine name
method if desired. As such, remove the confusion by axing the value.