Allow multiple type conversions through a pipeline.
Ensure that an request to change format causes engines to be chosen
again. This is rather important for format specific engines such as
Gifsicle - as demonstrated by the added test, resizing a GIF should
use the specific engine to avoid performing an operation with one
that is less capable for the format.