Tweak gifsicle argument output so the command line concat works.
The Gifsicle engine may require multiple calls of the command line
utility in the case of a resize followed by an extract. To handle
this case but keep an arguments array that is a single level deep
(to allow is being concatenated if an error occurs) opt to delimit
the change with a semicolon.