Remove string args to createPipeline() being parsed to operations.
Make use of the fact that a query string is parsed to a series of
operations and simply enforce that the API going forward is to pass
the resulting operations array to cratePipeline(). This change is a
pre-requisite for starting to separate the query string parsing from
the core of the library.