Bring in the legacy query string parsing as a queryString function.
The legacy parser is closely coupled to the impro parser as internally
it functions as a preparation step which is then handed to the impro
query string parser.

Import it so it becomes another function exposed by the queryString
file - this allows any project that exposes the legacy format to get
rid of any previous dependencies and instead use impro alone.