Define the test suite in terms of a Pipeline.
Register a Pipeline type with Unexpected and rework the tests as a
series of specific operations against them. This change means that
the tests are closer to what consumers of the external API deal with
(namely pipeline objects) and also leans on the runtime type checking
such that any mistakes with objects will be indicated early.