Stop relying on an index where a boolean flag will do.
The code is functionally equivalent but is a bit safer in the case
of an a sync throw while selecting set of operations on a particular
engine where we were faking a non -1 index to get the stream failure
behaviour. If the index into the set of streams were ever to be used
to grab some information from the used engine, the caller would need
to know it was not a safe dereference.