Strip out defaultEngineName and finish simplifying options handling.
The defaultEngineName value has had no effect on engine choices for
some time, and any engine preference is in the order of registered
engines that correspond to any relevant choices/information about
image types. Its default was also incorrect. Remove it.

With that gone, we further notice that after prior rationalisation
of options handling in pipelines that are handed a list of supported
option names and use them to pick applicable subset of values, there
is no longer any need to assign them to the Impro object. Take the
opportunity to kill that stuff too.