Harden stream error handling for cases of a non-Error instance emit.
We have received an error report that suggests a stream registered
with a Pipeline emitted a non-Error (specifically undefined). This
really should not happen, but since in principle it was intended
that arbitrary streams could be registered, we have always opted
to be extremely thorough around error handling. Opt to make sure
such an occurrence results in gracefully tearing down the stream
and attempt to provide some breadcrumbs around what occurred.