Introduce StdinoutStream.
This commit adds an implementation of a transform stream intended to
abstract spawning a command line utility which receives data on stdin
and outputs a result on stdout. Any data written to the stream (i.e.
pipe()'ing a Readable stream to it or otherwise) is sent onwards to
the child process and any data it produces it available to be read
downstream. To the extent possible this is fully streamed and care
has been taken to handle various error conditions and non-determistic
execution ordering e.g. child process stdout versus its exit event.

The genesis for doing this came from looking into a report of an
uncaught exception related to the use of gifsicle. A code inspection
combined with having a fair share of "experences" with streams left
a somewhat uneasy feeling and, given that some better facilities are
available in newer node together (e.g. Transform streams) and with
an awareness behaviours have been subtlely tweaked, lead to concluding
it worthwhile to invest in using a new hardened wrapper going forward.
This culmination of that work built out using gifsicle as the target.

This is intended for potentially sharing across tools going forward.