Integrate the StdinoutStream wrapped with gifsicle engine.
Use the child process stdin/stdout stream abstraction added in the
previous commit to replace gifsicle-stream. Add additional tests to
cover the situations of a) an empty image b) a non-GIF image being
piped through the impro with the gifsicle engine selected and assert
that errors are correctly emitted.

A couple of notable side benefits of this work include a small overall
reduction in the depdendency count (thus installation size), unblocking
the use of a newer version of the underlying image library and debug
logging throughout the gifsicle engine controlled by the DEBUG env var.