Prioritise npm package gifsicle above the system version.
It occurred to me that the other ordering would leave the user with
no way to choose to use the package _despite_ a system binary being
available - this way the explicit inclusion of the package causes
its use regardless of the availability of a copy being on the path.

While here tweak the stdin/stdout stream test file to make it more
straightforward to check that fallback to a binary on the PATH works.