Fix application of the hide marker in the README by evaldown.
The hide marker is ignored with --inplace because, were it to attempt
to hide in place, that would result in removing source content of the
very document being degenrated. Sidestep this by placing a copy of
the README into a docs folder and have the repository level document
generated from that.

Note that as of this commit changes to README should be placed into
the docs folder and these will be folded in at package version time.